software/requirements.txt
spidev
RPi.GPIO
Pillow
waveshare-epaper
